refact: Delete Wrapper.tsx
뷰 컴포넌트와 스켈레톤을 분리시켰습니다.
