+++ src/utils.ts
@@ -1,5 +1,8 @@
 import qs from 'querystring'
 
+import { User, SemesterEnglish2Korean } from './types/models'
+import { GradePoint, gradeRangeAToC } from './types/dreamy'
+
 export const saveToken = () => {
   localStorage.setItem('token', qs.parse(window.location.search)[
     '?token'
@@ -10,10 +13,42 @@ export const tokenDelete = () => localStorage.removeItem('token')
 
 export const getToken = () => localStorage.getItem('token')
 
-export function count<T>(array: T[], itm: T) {
+function count<T>(array: T[], itm: T) {
   return array.filter(x => x === itm).length
 }
 
 const reduce = (pre: number, curr: number) => pre + curr
 
-export const sumArray = (arr: number[]) => arr.reduce(reduce)
+const sumArray = (arr: number[]) => arr.reduce(reduce)
+
+export type PostProcessor = ReturnType<typeof postProcesser>
+
+export const postProcesser = ({ name, semesters, averagePoint }: User) => {
+  const creadit = semesters
+    .map(semester => semester.totalCredit)
+    .reduce((acc: number, curr: number) => acc + curr)
+
+  const gradePoint = semesters
+    .map(({ subjects }) => subjects.map(({ grade }) => grade))
+    .flat()
+
+  const series = gradeRangeAToC.map(grade =>
+    count<GradePoint>(gradePoint, grade)
+  )
+
+  const gradeRate = [...series, gradePoint.length - sumArray(series)]
+
+  const semesterOmitOutside = semesters.filter(({ isOutside }) => !isOutside)
+
+  const gradeRateAverages = [
+    {
+      name,
+      data: semesterOmitOutside.map(semester => semester.averagePoint),
+    },
+  ]
+  const semesterNames = semesterOmitOutside.map(
+    ({ year, semester }) => `${year}년 ${SemesterEnglish2Korean[semester]}`
+  )
+
+  return { creadit, gradeRate, gradeRateAverages, semesterNames, averagePoint }
+}
